move video to lang
